requirements.txt
fastapi[standard]==0.113.0
pydantic==2.8.0
motor==3.6.0
python-dotenv==1.0.1
